Include RFC 2822 date, properly format raw emails
SpamAssasin will throw high warnings if the Date: header isn't included, or if the email address isn't properly formatted. While technically allowed in in the RFC, the out of the box SpamAssassin setup deployed on many servers penalizes emails that lack the <> around the To: header. See TO_NO_BRKTS_DIRECT.